pyproject.toml
[project]
name = "braintu-be"
version = "0.1.0"
description = ""
authors = [
    {name = "HarshitCodes2",email = "g1harshit7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (==5.0.6)",
    "djangorestframework (==3.15.2)",
    "python-dotenv (==1.0.1)",
    "djangorestframework-simplejwt (==5.3.1)",
    "psycopg2-binary (==2.9.9)"
]

[tool.poetry]
packages = [{include = "braintu_be"}]
package-mode = false

[tool.poetry.group.dev.dependencies]
django-stubs = "5.1.1"
djangorestframework-stubs = "3.15.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
